+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## vNEXT
+### Highlights :tada:
++ Support **optional attributes**, see [the docs](https://slinky.dev/docs/the-tag-api/) for more details [PR #275](https://github.com/shadaj/slinky/pull/275)
 
 ## [v0.6.1](https://slinky.dev)
 ### Highlights :tada: